Cache of Computationally Expensive Objects for Reuse

cache() creates and / or returns a 'BiocFileCache' object containing information about objects in the package-specific cache.

cache_exists() tests whether an object exists in the cache. The object is identified by the concatentation of the project_id and object_name argument.

cache_add() adds object to the cache, using project_id and object_name to create an identifier. The writer_function is used to save the object to the cache.

cache_read() retrieves an object from the cache, using project_id and object_name to create an identifier. The reader_function is used to read the object into R.

cache_remove() removes an object from the cache.

Usage

cache(create = TRUE)

cache_exists(project_id, object_name)

cache_add(
  object,
  project_id,
  object_name,
  writer_function = saveRDS,
  ...,
  overwrite = FALSE
)

cache_read(project_id, object_name, reader_function = readRDS, ...)

cache_remove(project_id, object_name)

Arguments

create

logical(1) create the cache if it does not exist? Default: TRUE.

project_id

character(1) project identifier, e.g., "3c9d586e-bd26-4b46-8690-3faaa18ccf38".

object_name

character(1) arbitrary object identifier. Cached objects created in Case Study vignettes are prefixed "local.", e.g., "local.countMatrix".

object

Any R object.

writer_function

a function taking as it's first argument object, a second argument the cache file path (determined internally) to store the object, and any additional arguments (... in cache_add()) required to write the object.

...

in cache_read(), ... are arguments passed to 'reader_function()`.

overwrite

logical(1) overwrite an existing object in the cache? With the default FALSE, attempts to write the object result in an error.

reader_function

a function taking as it's first argument the cache file path (determined internally) of the object, and any additional arguments (... in cache_add()) required to write the object.

Value

cache() returns a BiocFileCache object that can be manipulated using the interface defined in the BiocFileCache package.

cache_exists() returns a logical(1) value, TRUE when the object identified by projectId and object_name exists in the cache.

cache_add() returns, invisibly, the cache file path where the object is stored.

cache_read() returns the value returned by reader_function().

cache_remove() warns if the object is not in the cache, and returns FALSE invisibly, or removes the object and returns TRUE invisibly.

Examples

cache()
#> class: BiocFileCache
#> bfccache: /home/runner/.cache/R/hcaCaseStudies
#> bfccount: 0
#> For more information see: bfcinfo() or bfcquery()

faux_project_id <- "faux-project-id"
faux_object_name <- "local.fauxObject"

cache_exists(faux_project_id, faux_object_name)
#> [1] FALSE

cache_exists(faux_project_id, faux_object_name)       # FALSE
#> [1] FALSE
cache_add(letters, faux_project_id, faux_object_name)
cache_exists(faux_project_id, faux_object_name)       # TRUE
#> [1] TRUE

cache_read(faux_project_id, faux_object_name)
#>  [1] "a" "b" "c" "d" "e" "f" "g" "h" "i" "j" "k" "l" "m" "n" "o" "p" "q" "r" "s"
#> [20] "t" "u" "v" "w" "x" "y" "z"

cache_remove(faux_project_id, faux_object_name)
cache_exists(faux_project_id, faux_object_name)       # FALSE
#> [1] FALSE