Tests on Column Content

column_is_not_all_NA() tests whether all elements in column x are NA (undefined).

column_is_informative() returns TRUE when the column x contains values that differ across rows, and hence are informative about the status of the row.

column_is_experimentwide() returns TRUE when the column x is not all NA, and all values in the column are unique.

Usage

column_is_not_all_NA(x)

column_is_informative(x)

column_is_experimentwide(x)

Arguments

x

any R vector for which is.na() is defined.

Value

column_is_not_all_NA() returns TRUE if all elements in x are not NA, or FALSE otherwise.

Examples

example_tbl <- dplyr::tibble(
    x = letters,
    y = c(NA, tail(letters, -1)),
    z = rep(NA, length(letters)),
    w = "A"
)
example_tbl |>
    dplyr::select(where(column_is_not_all_NA))     # x, y, w
#> # A tibble: 26 × 3
#>    x     y     w    
#>    <chr> <chr> <chr>
#>  1 a     NA    A    
#>  2 b     b     A    
#>  3 c     c     A    
#>  4 d     d     A    
#>  5 e     e     A    
#>  6 f     f     A    
#>  7 g     g     A    
#>  8 h     h     A    
#>  9 i     i     A    
#> 10 j     j     A    
#> # … with 16 more rows

example_tbl |>
    dplyr::select(where(column_is_informative))    # x, y
#> # A tibble: 26 × 2
#>    x     y    
#>    <chr> <chr>
#>  1 a     NA   
#>  2 b     b    
#>  3 c     c    
#>  4 d     d    
#>  5 e     e    
#>  6 f     f    
#>  7 g     g    
#>  8 h     h    
#>  9 i     i    
#> 10 j     j    
#> # … with 16 more rows

example_tbl |>
    dplyr::select(where(column_is_experimentwide)) # w
#> # A tibble: 26 × 1
#>    w    
#>    <chr>
#>  1 A    
#>  2 A    
#>  3 A    
#>  4 A    
#>  5 A    
#>  6 A    
#>  7 A    
#>  8 A    
#>  9 A    
#> 10 A    
#> # … with 16 more rows