Summarize 'MatrixMart' files

mtx_count_matrix() parses one or several MatrixMart files with identical dimensions into a single dgCMatrix object.

Usage

mtx_count_matrix(
  .data,
  mtx_path,
  genes_path = NULL,
  barcodes_path = NULL,
  sep = "\t"
)

Arguments

.data

tibble containing a column with the value of the mtx_path parameter and, if non-NULL, columns with the value of parameters genes_path and barcodes_path.

mtx_path

character() paths to MatrixMart files on disk.

genes_path

character() (optional) paths to files containing gene (row) labels (e.g., ENSEMBL gene identifiers). The row labels must be in the first column of the file; fields are separated by the parameter sep; there can be no column headers in the file.

barcodes_path

character() (optional) paths to files containing cell (column) labels (e.g., barcodes). The column labels must be in the first column of the file; fields are separated by the parameter sep; there can be no column headers in the file.

sep

character(1), default "\t" (tab). Field separator in genes_path and barcodes_path files.

Value

mtx_count_matrix() returns a dgCMatrix sparse matrix with rows (genes) equal to the unique genes in all input files, and columns (cells) equal to the total cell count in the MatrixMart files. Samples (columns) are in the order present in .data.