Project-related Queries

project_information() queries the HCA databasee for project title, description, contact, DOI, and publication URI.

project_title() returns the title of the project, cleaned to remove trailing trailing ..

print.project_information() formats the result of project_information() in a more legible manner.

Usage

project_information(project_id)

project_title(project_id)

# S3 method for project_information
print(x, ...)

Arguments

project_id

character(1) project identifier, e.g., "3c9d586e-bd26-4b46-8690-3faaa18ccf38".

x

an object of class project_information, the result of a call to project_information().

...

additional arguments, required to conform with the print generic but not used.

Value

project_information() returns a tibble with a single row, and columns containing information about the project. The tibble is of class project_information and is printed in an interactive session formatted so long columns, e.g., projectDescription, are more easily read.

project_title() returns a character(1) vector containing the project title.

print.project_information() is invoked automatically when the result of project_information() is displayed for it's side effect of displaying the object.

Examples

project_id <- "3c9d586e-bd26-4b46-8690-3faaa18ccf38"
project_information(project_id)
#> Title
#>   Highly Parallel Genome-wide Expression Profiling of Individual Cells
#>   Using Nanoliter Droplets
#> Contributors (unknown order; any role)
#>   Joshua R Sanes, Rahul Satija, Alex K Shalek, Anindita Basu, James
#>   Nemesh, Itay Tirosh, John J Trombetta, Aviv Regev, Ami Day, Evan Z
#>   Macosko, Karthik Shekhar, Allison R Bialas, Steven A McCarroll, David
#>   A Weitz, Nolan Kamitaki, Emily M Martersteck, Melissa Goldman
#> Description
#>   Cells, the basic units of biological structure and function, vary
#>   broadly in type and state. Single-cell genomics can characterize cell
#>   identity and function, but limitations of ease and scale have
#>   prevented its broad application. Here we describe Drop-seq, a
#>   strategy for quickly profiling thousands of individual cells by
#>   separating them into nanoliter-sized aqueous droplets, associating a
#>   different barcode with each cell's RNAs, and sequencing them all
#>   together. Drop-seq analyzes mRNA transcripts from thousands of
#>   individual cells simultaneously while remembering transcripts' cell
#>   of origin. We analyzed transcriptomes from 44,808 mouse retinal cells
#>   and identified 39 transcriptionally distinct cell populations,
#>   creating a molecular atlas of gene expression for known retinal cell
#>   classes and novel candidate cell subtypes. Drop-seq will accelerate
#>   biological discovery by enabling routine transcriptional profiling at
#>   single-cell resolution.
#> doi: 10.1016/j.cell.2015.05.002
#> url: https://www.cell.com/fulltext/S0092-8674(15)00549-8
#> project: https://data.humancellatlas.org/explore/projects/3c9d586e-bd26-4b46-8690-3faaa18ccf38

project_title(project_id)
#> [1] "Highly Parallel Genome-wide Expression Profiling of Individual Cells Using Nanoliter Droplets"