Templates For New Case Studies

use_template() creates a new R markdown file as a starting point for a case study. The file requires manual editing to customize for the specific project.

Usage

use_template(project_id, case_study_path, overwrite = FALSE)

Arguments

project_id

character(1) project identifier, e.g., "3c9d586e-bd26-4b46-8690-3faaa18ccf38".

case_study_path

character(1) path to the directory in which the case study will be created.

overwrite

logical(1) overwrite a case study created by this function with the same project_id and case_study_path? Default: FALSE.

Value

use_template() returns the path to the rendered template (in case_study_path) with approriate substitution of variables.

Details

use_template() queries the HCA data portal for the project title associated with project_id, and uses this for the title and vignette index entry of the rendered markdown doeucment. The project_id is injected into the document, so that references to project_id are defined. Rendering the template also injects the current date, package name, and package version.

Examples

project_id <- "3c9d586e-bd26-4b46-8690-3faaa18ccf38"
case_study_path <- tempfile(); dir.create(case_study_path)
rmd <- use_template(project_id, case_study_path)
readLines(rmd) |>
    head() |>
    noquote()
#> [1] ---                                                                                                               
#> [2] title: "Case Study: Highly Parallel Genome-wide Expression Profiling of Individual Cells Using Nanoliter Droplets"
#> [3] author:                                                                                                           
#> [4]   - name: Martin Morgan                                                                                           
#> [5]     affiliation:                                                                                                  
#> [6]     - Roswell Park Comprehensive Cancer Center